New version 1.3.2-SNAPSHOT (generated by MC 4.1.5)
git-svn-id: https://sselab.de/lab2/private/svn/MontiCore/trunk/mclang/cd4analysis@24520 3ddcf219-3c89-4d16-bbf6-790a9aa18b9b

+++ src/test/java/de/monticore/umlcd4a/cocos/ebnf/TypesTest.java
@@ -98,7 +98,7 @@ public class TypesTest extends AbstractCoCoTest {
      * @see de.se_rwth.commons.logging.Log#doGetErrorCount()
      */
     @Override
-    protected int doGetErrorCount() {
+    protected long doGetErrorCount() {
       // always prevents system exit
       return 0;
     }
